+++ routes/web.php
@@ -20,9 +20,10 @@ Route::group(['namespace' => 'Auth'], function () {
 Route::group(['middleware' => ['auth']], function () {
     Route::get('/', 'CollectionController@index');
 
-    Route::group(['prefix' => 'collection'], function () {
-        Route::get('sync', 'CollectionController@showSyncForm')->name('collection.syncForm');
-        Route::put('sync', 'CollectionController@sync')->name('collection.sync');
+    Route::group(['prefix' => 'sync'], function () {
+        Route::get('/', 'CollectionSyncController@index')->name('sync.index');
+        Route::get('fetch-release/{id}', 'CollectionSyncController@fetchRelease')->name('sync.fetch-release');
+        Route::put('sync', 'CollectionSyncController@sync')->name('sync.sync');
     });
     Route::resource('collection', 'CollectionController')->only(['index', 'show']);
